Implemented: started creating context variable for userData, which will be distributed around the app
