+++ src/App.js
@@ -5,20 +5,26 @@ import {
 } from "react-router-dom";
 import GlobalReset from "./assets/CSS Components/GlobalReset";
 import GlobalStyles from "./assets/CSS Components/GlobalStyles";
-import TopBar from "./components/TopBar/TopBar";
+import TopBar from "./components/shared/TopBar/TopBar";
 import SignupPage from "./components/pages/Signup/SignupPage";
 import HomePage from "./components/pages/Home/HomePage";
+import UserDataContext from "./contexts/userDataContext";
+import { useState } from "react";
 
 export default function App() {
+	const [userData, setUserData] = useState({ name: "", zipCode: "", streetNumber: 0, complement: "", phone: "", image: "", birthDate: "", gender: "", token: "" });
+
   return (
     <BrowserRouter>
-        <GlobalReset />
-		<GlobalStyles />
-		<TopBar />
-        <Routes>
-            <Route path="/" element={<HomePage />} />
-            <Route path="/signup" element={<SignupPage />} />
-        </Routes>
+    	<GlobalReset />
+    	<GlobalStyles />
+		<UserDataContext.Provider value = {{userData, setUserData}}>
+			<TopBar />
+			<Routes>
+				<Route path="/" element={<HomePage />} />
+				<Route path="/signup" element={<SignupPage />} />
+			</Routes>
+		</UserDataContext.Provider>
     </BrowserRouter>
   );
 }
