+++ src/components/pages/Home/HomePage.js
@@ -1,45 +1,34 @@
-import {
-    HomeContainer, MenuContainer, ItemBox, ItemBoxes, ItemTitle, ItemPrice, MenuTopBar, ScrollLeftButton, ScrollRightButton,
-} from './HomePageStyle';
-import { mockItens } from '../../../mocks/mockItens';
-import { numberToCurrency } from '../../../utils/currencyUtils';
-import { useRef } from 'react';
-import { scrollHorizontally } from './HomePageFunctions';
+import { HomeContainer } from './HomePageStyle';
+import Menu from '../../shared/HomePageMenu/HomePageMenu';
+import { useEffect, useState } from 'react';
+import {mockMenu} from "../../../mocks/mockItens";
 
 export default function HomePage() {
 
+    const [menus, setMenus] = useState(mockMenu);
+
+    useEffect(() => {
+        //getMenus();
+    },[]);
+
+    if (menus.length === 0) {
+        return (
+        <HomeContainer>
+
+        </HomeContainer>
+        );
+    }
+
     return (
         <HomeContainer>
-            <Menu />
-            <Menu />
-            <Menu />
-            <Menu />
-            <Menu />
+            {menus.map( ({title, forwardMessage, itens}, index) => (
+                <Menu
+                    key={index}
+                    title={title}
+                    forwardMessage={forwardMessage}
+                    itens = {itens}
+                />
+            ))}
         </HomeContainer>
     )
-}
-
-function Menu() {
-    const divMenuRef = useRef();
-
-
-    return (
-        <MenuContainer>
-            <MenuTopBar>
-                Até R$60,00
-                <button> {"Dar uma olhada!!"} </button>
-            </MenuTopBar>
-            <ScrollLeftButton onClick={ (e) => scrollHorizontally(e, divMenuRef, "left", 220) }/>
-            <ItemBoxes ref={divMenuRef}>
-            {mockItens.map(( {image, name, price}, index) => (
-                <ItemBox key={index} >
-                    <ItemTitle> {name} </ItemTitle>
-                    <img alt={name} src={image} />
-                    <ItemPrice> {numberToCurrency(price)} </ItemPrice>
-                </ItemBox>
-            ))}
-            </ItemBoxes>
-            <ScrollRightButton onClick={ (e) => scrollHorizontally(e, divMenuRef, "right", 220) } />
-        </MenuContainer>
-    );
 }
