+++ src/components/shared/HomePageMenu/HomePageMenuFunctions.js
@@ -1,0 +1,12 @@
+function scrollHorizontally(event, reference, side, amount) {
+    event.preventDefault();
+    if (side === "left") {
+        reference.current.scrollLeft -= amount;
+    } else {
+        reference.current.scrollLeft += amount;
+    }
+}
+
+export {
+    scrollHorizontally,
+}
