+++ src/contexts/userDataContext.js
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+const UserDataContext = createContext();
+
+export default UserDataContext;
